+++ downloader/Extractors/Drtuber.java
@@ -34,7 +34,7 @@ import org.jsoup.select.Elements;
  * @author christopher
  */
 public class Drtuber extends GenericExtractor implements Searchable{
-    private static final int SKIP = 4;
+    private static final byte SKIP = 4;
     
     public Drtuber() { //this contructor is used for when you jus want to search
