+++ downloader/Extractors/Eporner.java
@@ -25,7 +25,7 @@ import org.jsoup.select.Elements;
  * @author christopher
  */
 public class Eporner extends GenericExtractor{
-    private static final int SKIP = 6;
+    private static final byte SKIP = 6;
     
     public Eporner() { //this contructor is used for when you jus want to search
         //https://eporner.com/hd-porn/IF2uT0kcojN/Blonde-Lass-Delicate-Hands/
