+++ downloader/Extractors/Ghettotube.java
@@ -32,7 +32,7 @@ import org.jsoup.select.Elements;
  * @author christopher
  */
 public class Ghettotube extends GenericExtractor implements Searchable{
-    private static final int SKIP = 3;
+    private static final byte SKIP = 3;
     
     public Ghettotube() { //this contructor is used for when you jus want to search
