+++ downloader/Extractors/Justporno.java
@@ -27,7 +27,7 @@ import org.jsoup.select.Elements;
  * @author christopher
  */
 public class Justporno extends GenericExtractor implements Searchable{
-    private static final int SKIP = 4;
+    private static final byte SKIP = 4;
     
     public Justporno() { //this contructor is used for when you jus want to search
