+++ downloader/Extractors/Tube8.java
@@ -34,7 +34,7 @@ import org.jsoup.select.Elements;
  * @author christopher
  */
 public class Tube8 extends GenericQueryExtractor implements Searchable{
-    private static final int SKIP = 4;
+    private static final byte SKIP = 4;
     
     public Tube8() { //this contructor is used for when you jus want to query
