+++ downloader/Extractors/Vidoza.java
@@ -28,7 +28,7 @@ import org.jsoup.nodes.Document;
  * @author christopher
  */
 public class Vidoza extends GenericExtractor {
-    private static final int SKIP = 3;
+    private static final byte SKIP = 3;
     
     public Vidoza() { //this contructor is used for when you jus want to search
