+++ downloader/Extractors/Vimeo.java
@@ -30,7 +30,7 @@ import org.jsoup.nodes.Document;
  * @author christopher
  */
 public class Vimeo extends GenericExtractor{
-    private static final int SKIP = 2;
+    private static final byte SKIP = 2;
     
     public Vimeo() { //this contructor is used for when you jus want to search
