+++ downloader/Extractors/Xhamster.java
@@ -36,7 +36,7 @@ import org.jsoup.select.Elements;
  * @author christopher
  */
 public class Xhamster extends GenericQueryExtractor implements Searchable{
-    private static final int SKIP = 5;
+    private static final byte SKIP = 5;
     
     public Xhamster() { //this contructor is used for when you jus want to query
