+++ downloader/Extractors/Xtube.java
@@ -28,7 +28,7 @@ import org.jsoup.select.Elements;
  * @author christopher
  */
 public class Xtube extends GenericExtractor implements Searchable{
-    private static final int SKIP = 6;
+    private static final byte SKIP = 6;
     
     public Xtube() { //this contructor is used for when you jus want to search
