+++ downloader/Extractors/Youjizz.java
@@ -28,7 +28,7 @@ import org.jsoup.select.Elements;
  * @author christopher
  */
 public class Youjizz extends GenericExtractor implements Searchable{
-    private static final int SKIP = 3;
+    private static final byte SKIP = 3;
 	
     public Youjizz() { //this contructor is used for when you jus want to search
